fix(letters): consolidate list() tests + require user param
Move 4 list-v2 tests from __tests__/ into co-located spec file, delete
the non-standard file, make the user param required on list() and fix
the call site in overdue.controller.ts to supply it.
